Changed to 16bits read on adc for node 2. Updated readme with some notes regarding singleton/multiton design pattern

+++ lib/adc_internal/adc_internal.h
@@ -13,7 +13,7 @@ class ADC_internal
             return instance;
         }
 
-        uint8_t read();
+        uint16_t read();
 
     private:
